+++ src/localisation/chineseSimplified.ts
@@ -506,6 +506,8 @@ const CHINESE_SIMPLIFIED: Translations = {
   [Phrase.ChatDeleteMessageTooltip]: '删除此聊天消息',
   [Phrase.ValidateLogPathLabel]: '验证日志路径',
   [Phrase.ValidateLogPathsDescription]: 'Warcraft Recorder 会检查你设置的日志路径是否指向受支持的 WoW 安装。你可以禁用此检查以允许设置不受支持的游戏模式。禁用此功能需自行承担风险。',
+  [Phrase.ValidateNtfsLabel]: '验证 NTFS',
+  [Phrase.ValidateNtfsDescription]: 'Warcraft Recorder 会检查日志路径是否位于 NTFS 文件系统上。您可以禁用此检查以允许设置不受支持的文件系统。禁用此功能需自行承担风险。',
   [Phrase.StartManualRecordingTooltip]: '开始手动录制',
   [Phrase.StopManualRecordingTooltip]: '停止手动录制',
   [Phrase.AdvancedCombatLoggingDisabledWarning]: '高级战斗日志未启用。请在魔兽世界中启用：系统 > 网络 > 高级战斗日志。注意：启用后需完全关闭魔兽世界 - 该设置仅在退出时写入 Config.wtf。',