+++ src/renderer/WindowsSettings.tsx
@@ -5,9 +5,11 @@ import { getLocalePhrase } from 'localisation/translations';
 import { setConfigValues } from './useSettings';
 import Switch from './components/Switch/Switch';
 import Label from './components/Label/Label';
-import { Dispatch, SetStateAction } from 'react';
+import { Dispatch, SetStateAction, useEffect, useRef } from 'react';
 import { Tooltip } from './components/Tooltip/Tooltip';
 import { Phrase } from 'localisation/phrases';
+
+const ipc = window.electron.ipcRenderer;
 
 interface IProps {
   appState: AppState;
@@ -17,12 +19,14 @@ interface IProps {
 
 const WindowsSettings = (props: IProps) => {
   const { appState, config, setConfig } = props;
-  const initialRender = React.useRef(true);
 
-  React.useEffect(() => {
+  const effect1 = useRef(false);
+  const effect2 = useRef(false);
+
+  useEffect(() => {
     // Don't fire on the initial render.
-    if (initialRender.current) {
-      initialRender.current = false;
+    if (!effect1.current) {
+      effect1.current = true;
       return;
     }
 
@@ -42,6 +46,18 @@ const WindowsSettings = (props: IProps) => {
     config.hideEmptyCategories,
     config.hardwareAcceleration,
   ]);
+
+  useEffect(() => {
+    // Seperate effect here as we need to reconfigure base after changing
+    // the validateNtfs option to avoid the user needing to restart.
+    if (!effect2.current) {
+      effect2.current = true;
+      return;
+    }
+
+    setConfigValues({ validateNtfs: config.validateNtfs });
+    ipc.reconfigureBase();
+  }, [config.validateNtfs]);
 
   const getSwitch = (
     preference: keyof ConfigurationSchema,
@@ -130,6 +146,15 @@ const WindowsSettings = (props: IProps) => {
     });
   };
 
+  const setValidateNtfs = (checked: boolean) => {
+    setConfig((prevState) => {
+      return {
+        ...prevState,
+        validateNtfs: checked,
+      };
+    });
+  };
+
   return (
     <div className="flex flex-row flex-wrap gap-x-8">
       {getSwitchForm('startUp', Phrase.RunOnStartupLabel, setRunOnStartup)}
@@ -153,6 +178,12 @@ const WindowsSettings = (props: IProps) => {
         Phrase.HardwareAccelerationLabel,
         setHardwareAcceleration,
         Phrase.HardwareAccelerationDescription,
+      )}
+      {getSwitchForm(
+        'validateNtfs',
+        Phrase.ValidateNtfsLabel,
+        setValidateNtfs,
+        Phrase.ValidateNtfsDescription,
       )}
     </div>
   );